Move functor_wrapper.h to include/rocksdb
